pyproject.toml
[project]
name = "beeai-agents"
version = "0.1.3"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "IBM Corp." }]
requires-python = ">=3.11,<4.0"
dependencies = [
    "acp-sdk", # version determined by beeai-sdk
    "beeai-sdk==0.0.11",
    "pydantic-settings>=2.7.1",
    "pydantic>=2.10.6",
    "pyyaml>=6.0.2",
    "beeai-framework==0.1.1",
]

[project.scripts]
beeai-agents = "beeai_agents.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
